Arrange gp well data relative to user supplied start_corner

Arrange gp well data relative to user supplied start_corner

Usage

arrange_by_dim(gp, dim)

Arguments

gp

A gp object

dim

Character. Either "row" or "col"

Value

A gp with gp$well_data arranged by dim and the non-dim (ie if dim = "row", the 'non-dim' is "col") relative to the start corner (start corner is at the top, farther from start corner is at bottom)