Turn a tidy data.frame into a gp

An opposite to serve()

Usage

gp_unserve(x, row = ".row", col = ".col", nrow = NULL, ncol = NULL)

Arguments

x

A tidy data.frame representation of plate data

row

Character. The column that represents the rows. Should be numeric, with the lowest number representing the topmost row.

col

Character. The name of the column that represents columns. Should be numeric, with the lowest number representing the leftmost column.

nrow

Optional numeric, denoting the number of rows in the plate. If not supplied, will be imputed from the largest number in the 'row' column

ncol

Optional numeric, denoting the number of columns in the plate. If not supplied, will be imputed from the largest number in the 'col' column

Value

a gp